Added refresh link to all home navigation types
This adds a refresh link to all homePager.jsp browse methods in a
consistent place, so that users can easily refresh the proper frame
without causing a whole page refresh that would interrupt the web
player.
This is a follow-on to #1339 which removed the auto-refresh feature.
A side effect of this cleanup was to change all spacing between sets
of elements to a standard 2em to the right which made the layout
less convoluted.