Add custom credentials AuthenticationProvider
The added class checks for incoming mechanism of auth (password-based or salt/token based) and authenticates accordingly.
It will first fetch and migrate legacy creds in the db for that user to appropriate types (so legacy undecided creds are automatically upgradeable in the future).
Based on the incoming mechanism type and the stored representation in the db, it then redirects to the appropriate encoder that checks whether the password matches.
Multiple creds means it checks all until if finds one that matches, thus the name.