Allow PUTs for JSP FORWARDs (for validation errors) in credsSettings
JSPs do not allow PUTs or DELETEs (for non-error pages). Even if the request is POSTed using the hidden field method, the HiddenFilter converts the request to a PUT based on the field, and then on a forward (say on a validation error), the JSP sees the request as a PUT and throws a 405.

Two approaches exist
- Declare the page as an error page: this is done here.
- Add a filter that intercepts FORWARDS and changes the method to a GET

Documentation for these approaches are:
- https://stackoverflow.com/questions/24673041/405-jsp-error-with-put-method
- https://github.com/spring-projects/spring-framework/issues/17446
- https://stackoverflow.com/questions/23886941/http-status-405-jsps-only-permit-get-post-or-head

Another option is of course to change the PUT to a POST, but that option was discarded in preference for retaining some RESTful semantics.

This change is NOT strictly necessary considering the update creds command that does PUT does not really need a lot of overt validation, thus will rarely (if ever, and only if majorly wrong or something is manually being tinkered with) be forwarded with validation errors.