Jaudiotagger does not support APE
Jaudiotagger does not support APE but JaudiotaggerParser insists it does. There have been several attempts to get this fix implemented in the original Airsonic but somehow it never stuck. Jaudiotagger DOES NOT SUPPORT APE, using it anyway causes APE files to be registered with a duration of 0 which makes them unplayable. By removing APE from the list of supported formats the scanner uses FFmpegParser.java for the format which does work.