Support CUE-indexed tracks - "The Return of the CUE File"
This change adds support for cue-indexed files. Indexed tracks are added to the database using their relative position in the parent file and are played invidually by means of a "split" transcoding which is inserted into the transcoding chain. Base files - the files from which tracks are indexed - are shown or hidden depending on whether the 'Hide cue-indexed files' option (on the Media Folders settings page) is set or not. Cue-indexed tracks are represented in the database by their path and start position. The start position is stored in a new 'start_position' field in the database with a default value of '-1.0' - negative values indicate non-indexed files.

New settings:

 - Media Folders/Hide cue-indexed files (boolean): show (unset) or hide (set) cue-indexed base files
 - Transcoding/Split command (text): command used to split off section of a file for streaming, the default command `ffmpeg -ss %o -t %d -i %s -vcodec copy -acodec copy -f %f -` achieves this without transcoding (i.e. losless) by splitting on frame boundaries.

This is a continuation of the cue-support patch I submitted to the original airsonic (airsonic/airsonic#856) which never got merged. This version has been refactored for airsonic-advanced, rebased on the recently-merged relative paths PR (airsonic-advanced#659), put on display in the bottom of a locked filing cabinet stuck in a disused lavatory with a sign on the door saying ‘Beware of the Leopard" to be finally (?) resubmitted as a new PR.

Database changes:

 - add column `start_position` (double) default -1.0 (used to identify and access indexed tracks)
 - add column `index_path` varchar (used to store and identify indexed files)
 - drop unique index on path+folder_id
 - add unique index on path+folder_id+start_position